refactor: card types move  index to types
