+++ src/components/common/Card/BigCard.tsx
@@ -1,5 +1,5 @@
 import { ComponentPropsWithoutRef, ElementRef, forwardRef } from 'react'
-import { CardExpiry, CardNumber } from './index'
+import { CardExpiry, CardNumber } from './types'
 
 export type BigCardProps = ComponentPropsWithoutRef<'div'> & {
   number: CardNumber
