+++ src/components/common/Card/index.tsx
@@ -3,9 +3,6 @@ import { EmptyCard } from './EmptyCard'
 import { SmallCard } from './SmallCard'
 import { BigCard } from './BigCard'
 
-export type CardNumber = `${number} ${number} ${number} ${number}`
-export type CardExpiry = `${number}/${number}`
-
 export const Card = {
   New: NewCard,
   Empty: EmptyCard,
