+++ src/components/common/Card/types.ts
@@ -1,0 +1,2 @@
+export type CardNumber = `${number} ${number} ${number} ${number}`
+export type CardExpiry = `${number}/${number}`
